requirements.txt
feedparser==6.0.11
newspaper3k==0.2.8
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
anthropic==0.18.1
pytest==8.0.0
pytest-mock==3.12.0
ruff==0.2.1
